Move index file content into ./lib
IMO, keeping source code in a lib directory makes it easier to find within a project when the project root contains a lot of different config files and dotfiles, which can make spotting the `index.js` file tricky.
